Fixed Bug
There was a bug that was occurring where the program would crash when trying to print a file due to on when trying to output the file path it was interpreting that part as an int so it didn't want to concatenate the two together. Also, I updated readme.md to start a future road map.